+++ OceanDesignSystem/Controllers/Components SwiftUI/ScrollableTabSwiftUIViewController.swift
@@ -59,260 +59,3 @@ struct ScrollableTabSwiftUIViewController_Preview: PreviewProvider {
         }
     }
 }
-
-extension OceanSwiftUI {
-    // MARK: Parameter
-
-    public class ScrollableTabViewParameters: ObservableObject {
-        @Published public var tabs: [OceanSwiftUI.TabModel] {
-            didSet {
-                if !tabs.isEmpty {
-                    selectedIndex = 0
-                } else {
-                    selectedIndex = -1
-                }
-            }
-        }
-        @Published public var selectedIndex: Int {
-            didSet {
-
-            }
-        }
-        @Published public var padding: EdgeInsets
-        public var onTabSelected: (Int, OceanSwiftUI.TabModel) -> Void
-
-        public var topPositions: [CGFloat] = []
-
-        public init(tabs: [OceanSwiftUI.TabModel] = [],
-                    selectedIndex: Int = -1,
-                    padding: EdgeInsets = .init(top: 0,
-                                                leading: 0,
-                                                bottom: 0,
-                                                trailing: 0),
-                    onTabSelected: @escaping (Int, OceanSwiftUI.TabModel) -> Void = { _, _ in }) {
-            self.tabs = tabs
-            self.selectedIndex = selectedIndex
-            self.padding = padding
-            self.onTabSelected = onTabSelected
-        }
-    }
-
-    public struct ScrollableTabView: View {
-
-        // MARK: Properties for UIKit
-
-        public lazy var hostingController = UIHostingController(rootView: self)
-        public lazy var uiView = hostingController.getUIView()
-
-        // MARK: Builder
-
-        public typealias Builder = (ScrollableTabView) -> Void
-
-        // MARK: Properties
-
-        @ObservedObject public var parameters: ScrollableTabViewParameters
-
-        // MARK: Private properties
-
-        @State private var showTabAtIndex: Int = 0
-        @State private var showContentAtIndex: Int = 0
-
-//        @State private var showIndex: Int = 0
-        @State private var contentOffset: CGPoint = .zero
-
-        // MARK: Constructors
-
-        public init(parameters: ScrollableTabViewParameters = ScrollableTabViewParameters()) {
-            self.parameters = parameters
-        }
-
-        public init(builder: Builder) {
-            self.init()
-            builder(self)
-        }
-
-        // MARK: View SwiftUI
-
-        public var body: some View {
-            VStack {
-                if #available(iOS 14.0, *) {
-                    ScrollViewReader { scrollReader in
-                        ScrollView(.horizontal, showsIndicators: false) {
-                            VStack(spacing: 0) {
-                                HStack(spacing: 0) {
-                                    ForEach(0..<parameters.tabs.count, id: \.self) { index in
-                                        getTab(index: index)
-                                    }
-                                }
-                                .frame(height: 60)
-
-                                Divider()
-                            }
-                        }
-                        .onChange(of: showTabAtIndex) { _ in
-                            withAnimation {
-                                scrollReader.scrollTo(showTabAtIndex, anchor: .leading)
-                            }
-                        }
-                        .onChange(of: contentOffset) { offset in
-                            var lastIndex = -1
-                            parameters.topPositions.forEach { position in
-                                if position < offset.y {
-                                    lastIndex += 1
-                                }
-                            }
-
-                            if lastIndex > -1, lastIndex != showTabAtIndex {
-                                showTabAtIndex = lastIndex
-                                parameters.selectedIndex = lastIndex
-                            }
-                        }
-                    }
-
-                    ScrollViewReader { scrollReader in
-                        OffsetObservingScrollView(offset: $contentOffset) {
-                            ForEach(0..<parameters.tabs.count, id: \.self) { index in
-                                getView(index: index)
-                                    .onTapGesture {
-                                        parameters.selectedIndex = index
-                                        showTabAtIndex = index
-                                    }
-                            }
-                        }
-                        .onChange(of: showContentAtIndex) { _ in
-                            withAnimation {
-                                scrollReader.scrollTo(showContentAtIndex, anchor: .top)
-                            }
-                        }
-                    }
-                }
-            }
-            .padding(parameters.padding)
-            .background(Color(Ocean.color.colorInterfaceLightPure))
-        }
-
-        @available(iOS 14.0, *)
-        @ViewBuilder
-        private func getTab(index: Int) -> some View {
-            VStack {
-                Spacer()
-                HStack {
-                    OceanSwiftUI.Typography.heading4 { view in
-                        view.parameters.text = parameters.tabs[index].title
-                        view.parameters.lineLimit = 0
-                        view.parameters.textColor = parameters.selectedIndex == index
-                        ? Ocean.color.colorBrandPrimaryPure
-                        : Ocean.color.colorInterfaceDarkUp
-                    }
-
-                    if let badgeNumber = parameters.tabs[index].badgeNumber {
-                        OceanSwiftUI.Badge { badge in
-                            badge.parameters.count = badgeNumber
-                            badge.parameters.status = parameters.selectedIndex == index
-                            ? parameters.tabs[index].badgeStatus
-                            : .disabled
-                            badge.parameters.size = .small
-                        }
-                    }
-                }
-                .padding(.horizontal, Ocean.size.spacingStackXs)
-                Spacer()
-
-                if parameters.selectedIndex == index {
-                    Color(Ocean.color.colorBrandPrimaryPure).frame(height: 2)
-                } else {
-                    Color.clear.frame(height: 2)
-                }
-            }
-            .animation(.default, value: parameters.selectedIndex)
-            .onTapGesture {
-                showContentAtIndex = index
-                parameters.selectedIndex = index
-                parameters.onTabSelected(index, parameters.tabs[index])
-            }
-            .id(index)
-        }
-
-        @ViewBuilder
-        private func getView(index: Int) -> some View {
-            VStack(spacing: 0) {
-                GeometryReader { geometryReader in
-                    Rectangle()
-                        .frame(height: 0)
-                        .onAppear {
-                            let frame = geometryReader.frame(in: .named(UUID())).minY
-                            print("SWPOCAS -> Index: \(frame)")
-                            parameters.topPositions.append(frame - 138)
-                        }
-                }
-                .frame(height: 0)
-
-                parameters.tabs[index].contentView
-                    .id(index)
-            }
-        }
-
-
-    }
-}
-
-struct PositionObservingView<Content: View>: View {
-    var coordinateSpace: CoordinateSpace
-    @Binding var position: CGPoint
-    @ViewBuilder var content: () -> Content
-
-    var body: some View {
-        content()
-            .background(GeometryReader { geometry in
-                Color.clear.preference(
-                    key: PreferenceKey.self,
-                    value: geometry.frame(in: coordinateSpace).origin
-                )
-            })
-            .onPreferenceChange(PreferenceKey.self) { position in
-                self.position = position
-            }
-    }
-}
-
-private extension PositionObservingView {
-    struct PreferenceKey: SwiftUI.PreferenceKey {
-        static var defaultValue: CGPoint { .zero }
-
-        static func reduce(value: inout CGPoint, nextValue: () -> CGPoint) {
-            // No-op
-        }
-    }
-}
-
-struct OffsetObservingScrollView<Content: View>: View {
-    var axes: Axis.Set = [.vertical]
-    var showsIndicators = true
-    @Binding var offset: CGPoint
-    @ViewBuilder var content: () -> Content
-
-    // The name of our coordinate space doesn't have to be
-    // stable between view updates (it just needs to be
-    // consistent within this view), so we'll simply use a
-    // plain UUID for it:
-    private let coordinateSpaceName = UUID()
-
-    var body: some View {
-        ScrollView(axes, showsIndicators: showsIndicators) {
-            PositionObservingView(
-                coordinateSpace: .named(coordinateSpaceName),
-                position: Binding(
-                    get: { offset },
-                    set: { newOffset in
-                        offset = CGPoint(
-                            x: -newOffset.x,
-                            y: -newOffset.y
-                        )
-                    }
-                ),
-                content: content
-            )
-        }
-        .coordinateSpace(name: coordinateSpaceName)
-    }
-}
